+++ slim/base/_view/base_view.py
@@ -2,6 +2,7 @@ import asyncio
 import json
 import logging
 import time
+from collections import Mapping
 from types import FunctionType
 from typing import Set, List, Union, Optional, Dict
 from unittest import mock
@@ -9,7 +10,7 @@ from unittest import mock
 from aiohttp import hdrs, web
 from ipaddress import IPv4Address, IPv6Address, ip_address
 
-from aiohttp.web_request import BaseRequest
+from aiohttp.web_request import BaseRequest, FileField
 from multidict import MultiDict, CIMultiDictProxy
 
 from slim import Application, json_ex_dumps
@@ -17,7 +18,7 @@ from slim.base.helper import create_signed_value, decode_signed_value
 from slim.base.permission import Permissions
 from slim.base.types.temp_storage import TempStorage
 from slim.base.user import BaseUser, BaseUserViewMixin
-from slim.exception import NoUserViewMixinException
+from slim.exception import NoUserViewMixinException, InvalidPostData
 from slim.retcode import RETCODE
 
 from slim.utils import MetaClassForInit, async_call, sentinel, sync_call
@@ -90,7 +91,7 @@ class BaseView(metaclass=MetaClassForInit):
 
     @classmethod
     def use_lst(cls, name, summary=None, va_query=None, va_post=None, va_headers=None, va_resp=None, deprecated=False):
-        return cls.use_lst(name, summary=summary, va_query=va_query, va_post=va_post,
+        return cls._use_lst(name, summary=summary, va_query=va_query, va_post=va_post,
                            va_headers=va_headers, va_resp=va_resp, deprecated=deprecated)
 
     @classmethod
@@ -283,7 +284,10 @@ class BaseView(metaclass=MetaClassForInit):
             self._params_cache = MultiDict(self._request.query)
         return self._params_cache
 
-    async def post_data(self) -> "Optional[MultiDict[Union[str, bytes, FileField]]]":
+    async def post_data(self) -> "Optional[Mapping[str, Union[str, bytes, FileField]]]":
+        """
+        :return: 在有post的情况下必返回Mapping，否则返回None
+        """
         if self.method not in BaseRequest.POST_METHODS:
             return
 
@@ -292,8 +296,10 @@ class BaseView(metaclass=MetaClassForInit):
         if self._request.content_type == 'application/json':
             try:
                 self._post_data_cache = await self._request.json()
+                if not isinstance(self._post_data_cache, Mapping):
+                    raise InvalidPostData('post data should be a mapping.')
             except json.JSONDecodeError:
-                self._post_data_cache = None
+                self._post_data_cache = {}
         else:
             # post body: form data
             self._post_data_cache = await self._request.post()