+++ tests/utils_tests/test_utils_pagination.py
@@ -9,7 +9,7 @@ def test_pagination():
 
     assert pg['first_page'] is None
     assert pg['last_page'] == 100
-    assert pg['page_numbers'] == [1,2,3,4,5]
+    assert pg['numbers'] == [1, 2, 3, 4, 5]
     assert pg['info']['page_size'] == 10
     assert pg['info']['page_count'] == 100
     assert pg['info']['items_count'] == 1000
@@ -21,7 +21,7 @@ def test_pagination():
 
     assert pg['first_page'] is None
     assert pg['last_page'] == 100
-    assert pg['page_numbers'] == [1,2,3,4,5]
+    assert pg['numbers'] == [1, 2, 3, 4, 5]
     assert pg['info']['page_size'] == 10
     assert pg['info']['page_count'] == 100
     assert pg['info']['items_count'] == 1000
@@ -33,7 +33,7 @@ def test_pagination():
 
     assert pg['first_page'] == 1
     assert pg['last_page'] == 100
-    assert pg['page_numbers'] == [2,3,4,5,6]
+    assert pg['numbers'] == [2, 3, 4, 5, 6]
     assert pg['info']['page_size'] == 10
     assert pg['info']['page_count'] == 100
     assert pg['info']['items_count'] == 1000